feat(uip-preview): add preview plugin smooth content change animation
