+++ src/core/preview/preview.less
@@ -1,16 +1,24 @@
 .uip-preview {
-  display: block;
+  display: flex;
+  align-items: flex-start;
   grid-area: preview;
 
   &-inner {
+    flex: 1 0 auto;
     height: 100%;
     max-width: 100%;
     max-height: 100%;
     overflow: auto;
+  }
 
-    &.resizable {
-      resize: both;
-    }
+  &[resizable] &-inner {
+    resize: both;
+  }
+
+  &-stub {
+    flex: 0 0 0;
+    width: 0;
+    transition: height 0.5s linear;
   }
 
   &.centered-content {
